// api/src/domains/auth/core/schemas/register.schema.ts
import { Schema } from '@effect/schema';

export const RegisterInput = Schema.Struct({
  name: Schema.String.pipe(Schema.minLength(2)),
  email: Schema.String.pipe(Schema.pattern(/^[^\s@]+@[^\s@]+\.[^\s@]+$/)),
  password: Schema.String.pipe(Schema.minLength(8)),
});

export type RegisterInput = typeof RegisterInput.Type;